+++ src/main/java/roomescape/service/auth/AuthService.java
@@ -1,0 +1,8 @@
+package roomescape.service.auth;
+
+public interface AuthService {
+
+    String authenticate(AuthenticationRequest request);
+
+    AuthenticatedMemberProfile authorize(String token);
+}
