+++ src/main/java/roomescape/service/auth/AuthenticatedMemberProfile.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure.authentication;
+package roomescape.service.auth;
 
 public record AuthenticatedMemberProfile(Long id, String name, boolean isAdmin) {
 }
